@@ -40,11 +40,11 @@ public abstract class BasePreferenceActivity extends PreferenceActivity {
 	
 	@Override protected void onStart() {
 		super.onStart();
-	    EasyTracker.getInstance().activityStart(this);
+	    EasyTracker.getInstance(this).activityStart(this);
 	}
 	
 	@Override protected void onStop() {
 		super.onStop();
-		EasyTracker.getInstance().activityStop(this);
+		EasyTracker.getInstance(this).activityStop(this);
 	}
 }
